ImageInPainting/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
analytics-python==1.4.0
antlr4-python3-runtime==4.9.3
anyio==3.6.1
asttokens==2.0.5
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
backoff==1.10.0
bcrypt==4.0.0
brotlipy==0.7.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
coloredlogs==15.0.1
cycler==0.11.0
debugpy==1.5.1
diffusers @ https://github.com/huggingface/diffusers/archive/main.zip
entrypoints==0.3
executing==0.8.2
fastapi==0.81.0
ffmpy==0.3.0
filelock==3.8.0
flatbuffers==2.0.7
fonttools==4.37.1
frozenlist==1.3.1
fsspec==2022.8.2
gradio==3.2
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
huggingface-hub==0.9.1
humanfriendly==10.0
idna==3.4
importlib-metadata==4.12.0
invisible-watermark==0.1.5
Jinja2==3.1.2
jupyter-client==7.1.2
jupyter-core==4.9.1
kiwisolver==1.4.4
linkify-it-py==1.0.3
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.3
mdit-py-plugins==0.3.0
mdurl==0.1.2
mkl-fft==1.3.1
mkl-service==2.4.0
monotonic==1.6
mpmath==1.2.1
multidict==6.0.2
mypy-extensions==0.4.3
nest-asyncio==1.5.4
numpy==1.23.3
omegaconf==2.2.3
onnx==1.12.0
onnxruntime==1.12.1
opencv-python==4.6.0.66
orjson==3.8.0
packaging==21.3
pandas==1.4.4
paramiko==2.11.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.4.1
prompt-toolkit==3.0.26
protobuf==3.20.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycryptodome==3.15.0
pydantic==1.10.1
pydub==0.25.1
Pygments==2.11.2
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.2.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
regex==2022.9.13
requests==2.28.1
rfc3986==1.5.0
sniffio==1.3.0
stack-data==0.1.4
starlette==0.19.1
sympy==1.11.1
tokenizers==0.12.1
torch==1.12.1
torchvision==0.13.1
tornado==6.1
tqdm==4.64.1
traitlets==5.1.1
transformers==4.19.2
typing_extensions==4.3.0
uc-micro-py==1.0.1
urllib3==1.26.12
uvicorn==0.18.3
wcwidth==0.2.5
websockets==10.3
yarl==1.8.1
zipp==3.8.1
